refactor(Filters): move patterns and rename validate
